refactor: Move authentication logic to middleware and clean up main application file
